+++ core/server/api/canary/utils/serializers/output/mappers/actions.js
@@ -1,0 +1,7 @@
+const clean = require('../utils/clean');
+
+module.exports = (model, frame) => {
+    const attrs = model.toJSON(frame.options);
+    clean.action(attrs);
+    return attrs;
+};
