+++ core/server/api/canary/utils/serializers/output/mappers/images.js
@@ -1,0 +1,5 @@
+const url = require('../utils/url');
+
+module.exports = (path) => {
+    return url.forImage(path);
+};
