+++ core/server/api/canary/utils/serializers/output/mappers/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+    actions: require('./actions'),
+    emails: require('./emails'),
+    images: require('./images'),
+    integrations: require('./integrations'),
+    labels: require('./labels'),
+    pages: require('./pages'),
+    posts: require('./posts'),
+    settings: require('./settings'),
+    tags: require('./tags'),
+    users: require('./users')
+};
