+++ core/server/api/canary/utils/serializers/output/mappers/labels.js
@@ -1,0 +1,4 @@
+module.exports = (model, frame) => {
+    const jsonModel = model.toJSON ? model.toJSON(frame.options) : model;
+    return jsonModel;
+};
