+++ core/server/api/canary/utils/serializers/output/mappers/tags.js
@@ -1,0 +1,11 @@
+const clean = require('../utils/clean');
+const url = require('../utils/url');
+
+module.exports = (model, frame) => {
+    const jsonModel = model.toJSON ? model.toJSON(frame.options) : model;
+
+    url.forTag(model.id, jsonModel, frame.options);
+    clean.tag(jsonModel, frame);
+
+    return jsonModel;
+};
