+++ core/server/api/canary/utils/serializers/output/mappers/users.js
@@ -1,0 +1,12 @@
+const clean = require('../utils/clean');
+const url = require('../utils/url');
+
+module.exports = (model, frame) => {
+    const jsonModel = model.toJSON ? model.toJSON(frame.options) : model;
+
+    url.forUser(model.id, jsonModel, frame.options);
+
+    clean.author(jsonModel, frame);
+
+    return jsonModel;
+};
